ACT&INT: make "Show macro expansion" available for custom derive
